// src/types/trello.ts
export interface TrelloBoard {
  id: string;
  name: string;
}

export interface TrelloList {
  id: string;
  name: string;
}

export interface TrelloLabel {
  id: string;
  name: string;
}

export interface TrelloCardResponse {
  id: string;
}

export interface TrelloChecklistResponse {
  id: string;
}

export interface TrelloMember {
  id: string;
}